Moved is_db_uri to top-level.
