+++ src/pacsanini/utils.py
@@ -5,6 +5,8 @@
 """Simple utilities to facilitate the ingestion of resource values
 into the application.
 """
+import re
+
 from typing import List
 
 import pandas as pd
@@ -56,3 +58,23 @@ def read_resources(resources_path: str, query_level: QueryLevel) -> List[str]:
             resources = resources["StudyInstanceUID"].unique().tolist()
 
     return resources
+
+
+SUPPORTED_DB_DIALECTS = [
+    re.compile(r"postgresql(\+[\w\d]+)?://"),
+    re.compile(r"mysql(\+[\w\d]+)?://"),
+    re.compile(r"mariadb(\+[\w\d]+)?://"),
+    re.compile(r"oracle(\+[\w\d]+)?://"),
+    re.compile(r"sqlite://"),
+]
+
+
+def is_db_uri(uri: str) -> bool:
+    """Return true if the URI is for a known database. False
+    otherwise (eg: it is a file path).
+    """
+    uri_lower = uri.lower()
+    for dialect in SUPPORTED_DB_DIALECTS:
+        if dialect.match(uri_lower):
+            return True
+    return False
